Remove Event and Command types
While writing the use cases in the README, it became apparent that
having separate types does add any value in the current state since
there are other types that depend on the specific types. The semantics
of the message is defined by the domain-specific aspects such as the
type string and the data.

Signed-off-by: Byron Ruth <b@devel.io>